separate components and user object
