+++ my-react-app/src/App.jsx
@@ -1,12 +1,13 @@
-import Cred from './Cred.jsx'
+import React from "react";
+import Crud from "./Crud.jsx";
 
 function App() {
-    
-    return(
-      <>
-        <Cred/>
-      </>
-    );
+  return (
+    <>
+      <Crud />
+    </>
+  );
 }
 
-export default App
+export default App;
+
